backend/requirements.txt
# Snip Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
requests==2.31.0

# RAG Dependencies
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf==4.0.1
python-docx==1.1.0
tiktoken==0.5.2

# Utilities
aiofiles==23.2.1
uuid==1.30
